[core-amqp] make flaky tests not flaky
